+++ packages/react-start-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tanstack/react-start-client
 
+## 1.168.16
+
+### Patch Changes
+
+- Updated dependencies [[`e2dd204`](https://github.com/TanStack/router/commit/e2dd2049cb42eb219d3b447b8605066d19d9c1fa)]:
+  - @tanstack/router-core@1.171.15
+  - @tanstack/react-router@1.170.18
+  - @tanstack/start-client-core@1.170.14
+
 ## 1.168.15
 
 ### Patch Changes