+++ packages/react-start-rsc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/react-start-rsc
 
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [[`e2dd204`](https://github.com/TanStack/router/commit/e2dd2049cb42eb219d3b447b8605066d19d9c1fa)]:
+  - @tanstack/router-core@1.171.15
+  - @tanstack/react-router@1.170.18
+  - @tanstack/start-client-core@1.170.14
+  - @tanstack/start-plugin-core@1.171.20
+  - @tanstack/start-server-core@1.169.17
+  - @tanstack/start-storage-context@1.167.17
+
 ## 0.1.26
 
 ### Patch Changes