+++ packages/router-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/router-generator
 
+## 1.167.19
+
+### Patch Changes
+
+- Updated dependencies [[`e2dd204`](https://github.com/TanStack/router/commit/e2dd2049cb42eb219d3b447b8605066d19d9c1fa)]:
+  - @tanstack/router-core@1.171.15
+
 ## 1.167.18
 
 ### Patch Changes