+++ packages/solid-start-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tanstack/solid-start-server
 
+## 1.167.22
+
+### Patch Changes
+
+- Updated dependencies [[`e2dd204`](https://github.com/TanStack/router/commit/e2dd2049cb42eb219d3b447b8605066d19d9c1fa)]:
+  - @tanstack/router-core@1.171.15
+  - @tanstack/solid-router@1.170.18
+  - @tanstack/start-server-core@1.169.17
+
 ## 1.167.21
 
 ### Patch Changes